requirements.txt
Adafruit-Blinka
adafruit-circuitpython-busdevice
adafruit-circuitpython-motor
adafruit-circuitpython-pca9685
adafruit-circuitpython-register
adafruit-circuitpython-requests
adafruit-circuitpython-servokit
adafruit-circuitpython-typing
Adafruit-GPIO
Adafruit-PCA9685
Adafruit-PlatformDetect
Adafruit-PureIO
backcall
decorator
ipython
ipython-genutils
jedi
Jetson.GPIO
parso
pexpect
pickleshare
prompt-toolkit
ptyprocess
pyftdi
Pygments
pylibi2c
pyserial
python-dotenv
pyusb
RPLCD
six
smbus
smbus2
spidev
traitlets
typing_extensions
wcwidth
gymnasium
PyYAML==6.0.1
adafruit_servokit
numpy==1.19.4
opencv-python==4.9.0.80